# ebook-converter/src/app/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BookConv API",
    "description": "Ebook format conversion API powered by Calibre. Supports 28+ formats including EPUB, MOBI, AZW3, PDF, DOCX, TXT, FB2, LIT, RTF, and more.\n\n**Base URL:** `https://bookconv.com/api`",
    "version": "1.0.0",
    "contact": {
      "name": "BookConv Support",
      "url": "https://bookconv.com"
    }
  },
  "servers": [
    {
      "url": "https://bookconv.com/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Bearer token for authenticated users (optional). Free users are rate-limited to 20 requests/minute per IP. Pro users get priority processing."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string", "nullable": true }
        }
      }
    },
    "parameters": {
...
      "BatchIdPath": {
        "name": "batchId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" },
        "description": "The batch ID returned from the batch convert endpoint"
      }
    },
...
  },
  "security": [{ "ApiKeyAuth": [] }],
  "paths": {
...
    "/convert/batch/{batchId}/download": {
      "get": {
        "tags": ["Batch Conversion"],
        "summary": "Download batch results as ZIP",
        "description": "Download all successfully converted files from a batch as a single ZIP archive.",
        "operationId": "downloadBatchZip",
        "parameters": [{ "$ref": "#/components/parameters/BatchIdPath" }],
        "responses": {
          "200": {
            "description": "ZIP archive of converted files",
            "content": {
              "application/zip": {
                "schema": { "type": "string", "format": "binary" }
              }
            },
            "headers": {
              "Content-Disposition": { "schema": { "type": "string" }, "example": "attachment; filename=\"converted-a1b2c3d4.zip\"" },
              "X-Batch-Id": { "schema": { "type": "string" } },
              "X-Completed-Files": { "schema": { "type": "integer" } }
            }
          },
          "404": {
            "description": "Batch not found or no completed files",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "examples": {
                  "notFound": { "value": { "error": "Batch not found" } },
                  "noCompleted": { "value": { "error": "No completed files to download" } }
                }
              }
            }
          }
        }
      }
    },
...
    "/download": {
      "get": {
        "tags": ["Downloads"],
        "summary": "Download a converted file",
        "description": "Download a previously converted file using a storage key and expiration timestamp. Rate limited to 50 downloads per hour per IP.",
        "operationId": "downloadFile",
        "parameters": [
          {
            "name": "k",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Storage key returned from the result endpoint"
          },
          {
            "name": "expires",
            "in": "query",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Unix timestamp (ms) when the link expires"
          }
        ],
        "responses": {
          "200": {
            "description": "File downloaded successfully",
            "content": {
              "application/octet-stream": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "400": {
            "description": "Missing required parameters",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Missing required parameters: k, expires" }
              }
            }
          },
          "410": {
            "description": "Download link expired",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Download link has expired" }
              }
            }
          },
          "429": {
            "description": "Too many downloads (50/hour per IP)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Too many downloads. Please try again later." }
              }
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "File not found" }
              }
            }
          },
          "503": {
            "description": "Download service unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Download service unavailable" }
              }
            }
          }
        }
      }
    }
  }
}